# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Lantern Tarot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Reading {
  id: number;
  spreadId: string;
  spreadName: string;
  /** @nullable */
  question?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  synthesis?: string | null;
  deckMode: string;
  focus: string;
  feelings?: string[];
  cardNames?: string[];
  createdAt: string;
}

export interface DrawnCardRecord {
  cardId: string;
  cardName: string;
  positionName: string;
  reversed: boolean;
}

export interface ReadingDetail {
  id: number;
  spreadId: string;
  spreadName: string;
  /** @nullable */
  question?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  synthesis?: string | null;
  deckMode: string;
  focus: string;
  feelings?: string[];
  createdAt: string;
  cards: DrawnCardRecord[];
}

export interface DrawnCardInput {
  cardId: string;
  cardName: string;
  positionName: string;
  reversed: boolean;
}

export interface ReadingInput {
  spreadId: string;
  spreadName: string;
  question?: string;
  notes?: string;
  synthesis?: string;
  deckMode: string;
  focus: string;
  feelings?: string[];
  cards: DrawnCardInput[];
}

export interface InsightBar {
  label: string;
  value: number;
}

export interface Insights {
  total: number;
  topCards: InsightBar[];
  feelings: InsightBar[];
  suits: InsightBar[];
  orientations: InsightBar[];
  monthly: InsightBar[];
}

export interface Settings {
  theme: string;
}

export interface SettingsInput {
  theme?: string;
}

export interface CustomDeck {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  cardCount: number;
  createdAt: string;
}

export interface CustomCard {
  id: number;
  deckId: number;
  name: string;
  glyph: string;
  /** @nullable */
  imageUrl?: string | null;
  uprightMeaning?: string;
  reversedMeaning?: string;
  uprightKeywords?: string[];
  reversedKeywords?: string[];
  /** @nullable */
  reflection?: string | null;
  position: number;
}

export interface CustomDeckDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  cardCount: number;
  createdAt: string;
  cards: CustomCard[];
}

export interface CustomDeckInput {
  name: string;
  description?: string;
}

export interface CustomCardInput {
  name: string;
  glyph?: string;
  imageUrl?: string;
  uprightMeaning?: string;
  reversedMeaning?: string;
  uprightKeywords?: string[];
  reversedKeywords?: string[];
  reflection?: string;
}

export type ListReadingsParams = {
limit?: number;
offset?: number;
};