# lib/api-zod/src/generated/types/customCard.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Lantern Tarot API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomCard {
  id: number;
  deckId: number;
  name: string;
  glyph: string;
  /** @nullable */
  imageUrl?: string | null;
  uprightMeaning?: string;
  reversedMeaning?: string;
  uprightKeywords?: string[];
  reversedKeywords?: string[];
  /** @nullable */
  reflection?: string | null;
  position: number;
}