# lib/api-zod/src/generated/types/customCardInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Lantern Tarot API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomCardInput {
  name: string;
  glyph?: string;
  imageUrl?: string;
  uprightMeaning?: string;
  reversedMeaning?: string;
  uprightKeywords?: string[];
  reversedKeywords?: string[];
  reflection?: string;
}