# lib/api-zod/src/generated/types/customDeck.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Lantern Tarot API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomDeck {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  cardCount: number;
  createdAt: string;
}