# lib/api-zod/src/generated/types/customDeckDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Lantern Tarot API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomCard } from './customCard';

export interface CustomDeckDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  cardCount: number;
  createdAt: string;
  cards: CustomCard[];
}