# lib/api-zod/src/generated/types/drawnCardInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Lantern Tarot API
 * OpenAPI spec version: 0.1.0
 */

export interface DrawnCardInput {
  cardId: string;
  cardName: string;
  positionName: string;
  reversed: boolean;
}